Add files via upload
I created a bank account in java that let the user create an account by adding his name, the id of the account and the initial balance in the account. The user can now deposit or withdraw money from that account after going through a payment verification that Robert created. So I put the bank account that I have created and the payment verification that Robert created together.